# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Barber Slot Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops/{slug}/slots/{date}/{serviceId}:
    get:
      operationId: getAvailableSlots
      tags: [slots]
      summary: Get available time slots for a shop on a given date for a given service
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
          description: Date in YYYY-MM-DD format
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
          description: Service to book
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlotsResponse"

  # ── BOOKINGS ──────────────────────────────────────────────────────────
  /shops/{slug}/bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for a shop (barber dashboard) for today
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Slot unavailable or advance booking rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slot conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shops/{slug}/bookings/{bookingId}/verify-otp:
    post:
      operationId: verifyArrivalOtp
      tags: [bookings]
      summary: Barber verifies customer arrival OTP to activate slot
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArrivalOtpInput"
      responses:
        "200":
          description: OTP verified, slot activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shops/{slug}/bookings/{bookingId}/no-show:
    post:
      operationId: markNoShow
      tags: [bookings]
      summary: Mark a booking as no-show
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked no-show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /shops/{slug}/bookings/{bookingId}/complete:
    post:
      operationId: completeBooking
      tags: [bookings]
      summary: Mark booking as completed
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /shops/{slug}/timeline/{date}:
    get:
      operationId: getTimeline
      tags: [bookings]
      summary: Get visual chair timeline for a given date
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
          description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Timeline data per chair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineResponse"
  /shops/{slug}/activity:
    get:
      operationId: getRecentActivity
      tags: [bookings]
      summary: Get recent booking activity feed (last 10 items)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /shops/{slug}/revenue:
    get:
      operationId: getRevenueStats
      tags: [bookings]
      summary: Revenue stats by day for last 7 days
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: weeksAgo
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Revenue stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDay"

# ── COMPONENTS ────────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
...
    Service:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        price:
          type: integer
          description: Price in INR
        durationMinutes:
          type: integer
          description: Service duration in minutes (buffer added server-side)
        isActive:
          type: boolean
      required: [id, shopId, name, price, durationMinutes, isActive]
...
    AvailableSlotsResponse:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
        shopIsOpen:
          type: boolean
        shopIsPaused:
          type: boolean
      required: [date, slots, shopIsOpen, shopIsPaused]
    TimeSlot:
      type: object
      properties:
        time:
          type: string
          description: "HH:MM format"
        available:
          type: boolean
        availableChairs:
          type: integer
      required: [time, available, availableChairs]

    # Booking
    BookingInput:
      type: object
      required: [customerName, customerPhone, serviceId, slotDate, slotTime, paymentType]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        serviceId:
          type: integer
        slotDate:
          type: string
          format: date
        slotTime:
          type: string
          description: "HH:MM format"
        paymentType:
          type: string
          enum: [token, full]
          description: "token = pay 1 INR now + rest in person; full = pay full upfront"
    Booking:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        serviceId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        slotDate:
          type: string
        slotTime:
          type: string
        slotEndTime:
          type: string
        chairNumber:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, active, completed, no_show, cancelled]
        paymentType:
          type: string
          enum: [token, full]
        amountPaid:
          type: integer
        totalAmount:
          type: integer
        arrivalOtp:
          type: ["string", "null"]
        service:
          $ref: "#/components/schemas/Service"
        shopName:
          type: string
        shopCity:
          type: string
        serviceName:
          type: string
        createdAt:
          type: string
      required: [id, shopId, serviceId, customerName, customerPhone, slotDate, slotTime, slotEndTime, chairNumber, status, paymentType, amountPaid, totalAmount, createdAt]
    ArrivalOtpInput:
      type: object
      required: [otp]
      properties:
        otp:
          type: string

    # Timeline
    TimelineResponse:
      type: object
      properties:
        date:
          type: string
        chairs:
          type: array
          items:
            $ref: "#/components/schemas/ChairTimeline"
        openTime:
          type: string
        closeTime:
          type: string
      required: [date, chairs, openTime, closeTime]
    ChairTimeline:
      type: object
      properties:
        chairNumber:
          type: integer
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
      required: [chairNumber, bookings]

    # Activity & Revenue
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [new_booking, arrival, no_show, completed, cancelled]
        customerName:
          type: string
        serviceName:
          type: string
        slotTime:
          type: string
        createdAt:
          type: string
      required: [id, type, customerName, serviceName, slotTime, createdAt]
    RevenueDay:
      type: object
      properties:
        date:
          type: string
        bookings:
          type: integer
        revenue:
          type: integer
      required: [date, bookings, revenue]