# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Barber Slot Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  shopId: number;
  name: string;
  /** Price in INR */
  price: number;
  /** Service duration in minutes (buffer added server-side) */
  durationMinutes: number;
  isActive: boolean;
}